i separated public views from admin views

+++ app/Http/Controllers/User/PermissionController.php
@@ -8,6 +8,7 @@ use App\Models\User\Resource;
 use App\Models\User\AccessLevel;
 use App\Models\User\Permission;
 use Illuminate\Support\Facades\DB;
+use Validator;
 
 
 class PermissionController extends Controller
@@ -25,7 +26,7 @@ class PermissionController extends Controller
     {
         $permissions=Permission::with('getAction','getAccessLevel','getResource')->get(['access_level_id','resource_id','action_id'])->groupBy('access_level_id');
         
-        return view('user.permissions.index',compact('permissions'));
+        return view('user.permissions.administrator.index',compact('permissions'));
     }
 
     /**
@@ -37,7 +38,7 @@ class PermissionController extends Controller
     {
         $resources=Resource::with('getActions')->get(['id','title']);
         $accessLevels=AccessLevel::all('id','title');
-        return view('user.permissions.create',compact('resources','accessLevels'));
+        return view('user.permissions.administrator.create',compact('resources','accessLevels'));
     }
 
     /**
@@ -51,39 +52,55 @@ class PermissionController extends Controller
         $validatedData = $request->validate([
             'accessLevel' =>'required|int',
         ]);
+
         $resources=Resource::all('id','title');
-        foreach ($resources as $r) {
-            $title=$r->title;
-            $actions=$request->$title;
-            try {
-                DB::transaction(function () use ($actions,$request,$r) {
-                  if($actions){
-                     for ($i=0; $i <count($actions) ; $i++) { 
-                       $permission= new Permission;
-                       $permission->access_level_id = $request->accessLevel;
-                       $permission->resource_id=$r->id;
-                       $permission->action_id=$actions[$i];
-                       $permission->save();
+        for ($i=0; $i <$resources->count() ; $i++) { 
+           if ($request->has($resources[$i]['title'])) {
+               foreach ($resources as $r) {
+                $title=$r->title;
+                $actions=$request->$title;
+                try {
+                    DB::transaction(function () use ($actions,$request,$r) {
+                      if($actions){
+                         for ($i=0; $i <count($actions) ; $i++) { 
+                           $permission= new Permission;
+                           $permission->access_level_id = $request->accessLevel;
+                           $permission->resource_id=$r->id;
+                           $permission->action_id=$actions[$i];
+                           $permission->save();
+                       }
                    }
-               }
-           });
+               });
 
-            } catch (Exception $e) {
+                } catch (Exception $e) {
 
-                session()->flash('message.type', 'danger');
-                session()->flash('message.content', 'Erreur lors de l\'ajout!');
+                    session()->flash('message.type', 'danger');
+                    session()->flash('message.content', 'Erreur lors de l\'ajout!');
+                }
+                session()->flash('message.type', 'success');
+                session()->flash('message.content', 'permission ajouté avec succès!');
+
+
             }
-            session()->flash('message.type', 'success');
-            session()->flash('message.content', 'permission ajouté avec succès!');
 
-
+            if ($request->save_close) {
+               return redirect()->route('permissions.index');
+           }else{
+            return redirect()->route('permissions.create');
         }
-
-        if ($request->save_close) {
-           return redirect()->route('permissions.index');
-       }else{
-        return redirect()->route('permissions.create');
     }
+}
+$validator = Validator::make($request->actions=[], [
+    'actions' => 'required',
+]);
+
+if ($validator->fails()) {
+    return back()
+    ->withErrors($validator)
+    ->withInput();
+}
+
+
 }
 
 
@@ -108,7 +125,7 @@ class PermissionController extends Controller
     {
         $accessLevel=Accesslevel::find($id);
         $resources=Resource::with('getActions','getPermissions')->get(['id','title']);
-        return view('user.permissions.edit',compact('resources','accessLevel'));
+        return view('user.permissions.administrator.edit',compact('resources','accessLevel'));
     }
 
     /**
@@ -121,45 +138,58 @@ class PermissionController extends Controller
     public function update(Request $request, $id)
     {
         if ($request->update) {
+
             $resources=Resource::all('id','title');
-            $existingPermissions=Permission::where('access_level_id',$id)->get();
-            try {
-                DB::transaction(function () use ($request,$id,$resources,$existingPermissions) {
-                    for ($i=0; $i <count($existingPermissions) ; $i++) { 
-                        $existingPermissions[$i]->forceDelete();
-                    }
-                    foreach ($resources as $r) {
-                        $title=$r->title;
-                        $actions=$request->$title;
+            for ($i=0; $i <$resources->count() ; $i++) { 
+               if ($request->has($resources[$i]['title'])) {
+                $existingPermissions=Permission::where('access_level_id',$id)->get();
+                try {
+                    DB::transaction(function () use ($request,$id,$resources,$existingPermissions) {
+                        for ($i=0; $i <count($existingPermissions) ; $i++) { 
+                            $existingPermissions[$i]->forceDelete();
+                        }
+                        foreach ($resources as $r) {
+                            $title=$r->title;
+                            $actions=$request->$title;
 
-                        if($actions){
-                         for ($i=0; $i <count($actions) ; $i++) { 
-                           $permission= new Permission;
-                           $permission->access_level_id = $id;
-                           $permission->resource_id=$r->id;
-                           $permission->action_id=$actions[$i];
-                           $permission->save();
+                            for ($i=0; $i <count($actions) ; $i++) { 
+                               $permission= new Permission;
+                               $permission->access_level_id = $id;
+                               $permission->resource_id=$r->id;
+                               $permission->action_id=$actions[$i];
+                               $permission->save();
+                           }
                        }
-                   }
+                   });
 
-               }
-           });
+                } catch (Exception $e) {
 
-            } catch (Exception $e) {
+                    session()->flash('message.type', 'danger');
+                    session()->flash('message.content', 'Erreur lors de l\'ajout!');
+                }
+                session()->flash('message.type', 'success');
+                session()->flash('message.content', 'Permissions modifiée avec succès!');
+                return redirect()->route('permissions.index');
 
-                session()->flash('message.type', 'danger');
-                session()->flash('message.content', 'Erreur lors de l\'ajout!');
             }
-            session()->flash('message.type', 'success');
-            session()->flash('message.content', 'Permissions modifiée avec succès!');
-
-        }else{
-            session()->flash('message.type', 'danger');
-            session()->flash('message.content', 'Modification annulée!');
-
         }
-        return redirect()->route('permissions.index');
+        $validator = Validator::make($request->actions=[], [
+            'actions' => 'required',
+        ]);
+
+        if ($validator->fails()) {
+            return back()
+            ->withErrors($validator)
+            ->withInput();
+        }
+
+    }else{
+        session()->flash('message.type', 'danger');
+        session()->flash('message.content', 'Modification annulée!');
+
     }
+    return redirect()->route('permissions.index');
+}
 
     /**
      * Remove the specified resource from storage.
@@ -169,6 +199,6 @@ class PermissionController extends Controller
      */
     public function destroy($id)
     {
-      
+
     }
 }
